Changed landing sounds
Added metal landing sound FX, renamed old landing to land_dirt.ogg
renamed metal floor image to metal_block-32.png

added metal landing sound to the preload and put it as default for main
level ;)

I was having issues with changing the platform block sprites to the
metal one locally so didn't mess with that here.

Also my browser does not fully scale the background image to be the
width of the canvas - it stops with a black background near the right
edge of the screen. Not a big issue, though.