Vector Shelf (background part asset for Factory)
Photoshop lets you make 'smart objects' which can be modified
independently from the background asset, if I modify this specific
object elsewhere It will reflect upon a main background project in
photoshop so that's a nice way of making a modulair background. I've
also made a terrible looking forklift but it needs some more work.
meanwhile I figure I'll just upload this asset for use elsewhere, maybe
in a market-level as a larger asset, aside from a smaller shelf on a
factory floor (which is going to be the main background with this as an
asset in it)